refactor(sidebar): create SidebarHeader component to streamline sidebar layout and improve code organization
